requirements.txt
torch
opencv-python
torchinfo
torchvision
tqdm
pyyaml
matplotlib
